Add example project and Dockerfile for Node dev
This mirrors as close as possible an example project and `Dockerfile` like the one used for ruby.

It also updates the README.md in `/node` with details on building, running and developing using the container.

Finally we update the project README to cover the fact we now have a Node.js example.